requirements.txt
Django==1.9.6
django-cors-headers==1.1.0
django-crispy-forms==1.6.0
django-filter==0.13.0
django-rest-auth==0.7.0
djangorestframework==3.3.3
six==1.10.0
wsgiref==0.1.2
